refactor(auth): extrait le HTML du composant de validation dans un fichier dédié
- Externalise le template de `VerificationComponent` dans un fichier `verification.html` pour améliorer la modularité
- Intègre des utilitaires partagés (`startCooldownTimer`, `clearCooldownTimer`) dans `VerificationComponent` et `VerificationModalView`
- Simplifie la logique de cooldown en centralisant les appels via les utilitaires
- Améliore la structure et le style des templates pour une meilleure maintenabilité et lisibilité
