+++ src/app/features/auth/components/verification/verification.html
@@ -1,0 +1,88 @@
+<div
+  class="w-full max-w-sm mx-auto px-4 py-3 sm:max-w-md sm:px-6 sm:py-8 md:max-w-lg lg:max-w-xl xl:max-w-2xl"
+>
+  <div
+    class="bg-background border border-border rounded-lg p-4 shadow-sm sm:rounded-xl sm:p-8 md:shadow-lg"
+  >
+    <h2 class="text-2xl font-bold text-center mb-4 text-text sm:text-3xl sm:mb-8 md:text-4xl">
+      Validation de l'inscription
+    </h2>
+
+    <p class="text-center text-muted mb-6 text-sm sm:text-base">
+      Un code de validation à 6 chiffres a été envoyé à votre adresse email :
+      <br /><strong class="text-text">{{ email() }}</strong>
+    </p>
+
+    <form [formGroup]="verificationForm" (ngSubmit)="onSubmit()" class="space-y-4 sm:space-y-6">
+      <div class="space-y-1 sm:space-y-2">
+        <label for="verificationCode" class="block text-sm font-medium text-text sm:text-base">
+          Code de validation <span class="text-error">*</span>
+        </label>
+        <input
+          id="verificationCode"
+          type="text"
+          formControlName="verificationCode"
+          maxlength="6"
+          pattern="[0-9]{6}"
+          class="w-full px-3 py-2 border rounded-md bg-background text-text placeholder:text-muted focus:outline-none focus:ring-2 focus:ring-primary focus:border-primary transition-all duration-200 sm:px-4 sm:py-3 sm:rounded-lg text-center text-lg sm:text-xl font-mono tracking-widest"
+          [class.border-error]="isFieldInvalid('verificationCode')"
+          [class.border-input]="!isFieldInvalid('verificationCode')"
+          placeholder="123456"
+          autocomplete="one-time-code"
+        />
+        @if (isFieldInvalid('verificationCode')) {
+          <p class="text-error text-xs mt-1 sm:text-sm">
+            @if (verificationForm.get('verificationCode')?.errors?.['required']) {
+              Le code de validation est requis
+            } @else if (verificationForm.get('verificationCode')?.errors?.['pattern']) {
+              Le code doit être composé de 6 chiffres
+            }
+          </p>
+        }
+      </div>
+
+      <app-button
+        type="submit"
+        color="primary"
+        [disabled]="verificationForm.invalid"
+        [isLoading]="authService.isLoading()"
+        customClass="w-full py-3 text-base font-semibold mt-4 sm:py-4 sm:text-lg sm:mt-8"
+      >
+        @if (authService.isLoading()) {
+          Validation en cours...
+        } @else {
+          Valider
+        }
+      </app-button>
+    </form>
+
+    <div class="mt-4 text-center sm:mt-6">
+      <p class="text-muted text-sm sm:text-base mb-3">Vous n'avez pas reçu le code ?</p>
+
+      <button
+        type="button"
+        (click)="resendCode()"
+        [disabled]="authService.isLoading() || resendCooldown() > 0"
+        class="text-primary hover:text-primary-600 font-medium transition-colors duration-200 underline-offset-4 hover:underline disabled:opacity-50 disabled:cursor-not-allowed text-sm sm:text-base"
+      >
+        @if (resendCooldown() > 0) {
+          Renvoyer le code ({{ resendCooldown() }}s)
+        } @else if (authService.isLoading()) {
+          Envoi en cours...
+        } @else {
+          Renvoyer le code
+        }
+      </button>
+    </div>
+
+    <div class="mt-4 text-center sm:mt-6">
+      <button
+        type="button"
+        (click)="navigateToSignup()"
+        class="text-muted hover:text-text font-medium transition-colors duration-200 underline-offset-4 hover:underline text-sm sm:text-base"
+      >
+        Modifier les informations
+      </button>
+    </div>
+  </div>
+</div>
